Rename processModule to processModuleImports
