+++ javascript/processingFunctions/processSnippetHtml.js
@@ -7,7 +7,7 @@ import {
 } from "./warnings.js";
 import { chapterIndex } from "../parseXmlHtml";
 import recursiveProcessPureText from "./recursiveProcessPureText";
-import { processRuneModule } from "./processModule.js";
+import { processRuneModule } from "./processModuleImports.js";
 
 const snippetStore = {};
