+++ javascript/processingFunctions/processSnippetJs.js
@@ -4,7 +4,7 @@ import {
   repeatedNameWarning
 } from "./warnings.js";
 import recursiveProcessPureText from "./recursiveProcessPureText";
-import { processRuneModule } from "./processModule.js";
+import { processRuneModule } from "./processModuleImports.js";
 
 const snippetStore = {};
